Fix: Orbiter-Angriff konnte nur bis 255 gespeichert werden

+++ common/mysql_tables.php
@@ -233,7 +233,7 @@ $tables_add = array(
   `planetenGebPlanet` varchar(255) NOT NULL,
   `planetenGebOrbit` varchar(100) NOT NULL,
   `planetenGebSpezial` varchar(100) NOT NULL,
-  `planetenOrbiter` tinyint(3) unsigned NOT NULL DEFAULT 0,
+  `planetenOrbiter` smallint(5) unsigned NOT NULL DEFAULT 0,
   `planetenMyrigate` int(10) unsigned NOT NULL DEFAULT 0,
   `planetenRiss` int(10) unsigned NOT NULL DEFAULT 0,
   `planetenGateEntf` mediumint(8) unsigned DEFAULT NULL,
